requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
environs==9.3.3
idna==3.2
marshmallow==3.13.0
multidict==5.1.0
PyMySQL==1.0.2
python-dotenv==0.19.0
pytz==2021.1
typing-extensions==3.10.0.2
yarl==1.6.3
Pillow==8.1.0
pilmoji==2.0.1